+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add `removeNewLine` method
 - Add `capitalizeEverySentence` method
 - Add `contains` method
+- Add `capitalizeEachWords` method
 
 ## 0.0.1 (2024-09-02) - Not for Production